Fix incorrect parameter reference in plugin description
The description of the plugin in the 'influactive-modal-form-brochure.pot' was updated to replace the an incorrect parameter reference, from '?FILE_ID=ID' to '?file=ID'. This change improves the accuracy of the plugin documentation, ensuring users are provided with correct instructions. Additionally, the line references to the 'modal-form-brochure.php' file have been updated to match with the code updates.