Update plugin version to 1.3
The version number of the Influactive Modal Forms Brochure plugin was updated from 1.2 to 1.3 across several files including the plugin PHP file, the README files, and the language files (MO, PO, and POT). This update is vital to indicate the most recent plugin version and ensure compatibility with any version-dependent features or functionality.