Update Cart entity to include product variants
Barcode field added to the cart entity in order to allow for the recognition and handling of specific product variants. This change enables the cart to appropriately calculate and display quantities and prices according to specific product variations chosen by the user. Removed unnecessary code referencing the old product model.