Update form fields configuration documentation
Refactored the README.md and README.txt files for 'influactive-modal-form-brochure' to simplify form field configurations in the term of usability. Updated the instructions on how the forms could be configured, reducing the complexity of the setup process. This will make the process more intuitive for users and developers, enhancing usability of the module.