Update version to 1.5.1 and improve plugin dependency handling
The version of modal-form-brochure.php has been updated to 1.5.1 in several files including README.md, README.txt, and package.json. Also, enhanced the checking of the 'Forms everywhere by Influactive' plugin's availability. 'admin_init' action hook is now used to check if the plugin is active before running further code, preventing fatal errors previously encountered when the plugin was inactive. This improves the robustness and user-experience of the software.