Update package details and phpunit version
The package name and description in composer.json have been updated to more accurately reflect the plugin's functionality. The license has been updated to GPL-2.0-or-later, to align with open source licensing practices.

The phpunit version has also been updated to 10.3.1 in the composer.lock file. This update ensures that the package is tested against the latest version of PHPUnit for improved compatibility and security.